backend/requirements.txt
Flask==2.0.1
Flask-CORS==5.0.1
Flask-JWT-Extended==4.3.1
psycopg2-binary==2.9.1
redis==3.5.3
bcrypt==3.2.0
Werkzeug==2.0.3